// src/app/api/fn/warranty/warranty-claims-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedWarrantyClaimList } from '../../models/paginated-warranty-claim-list';

export interface WarrantyClaimsList$Params {

/**
 * true: received, sent for repair or ready for collection.
 */
  open?: 'false' | 'true';

/**
 * true: claimed after the warranty's end date.
 */
  out_of_warranty?: 'false' | 'true';

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Matches the warranty number, the product's code and name, and the customer's name and phone.
 */
  search?: string;

/**
 * Document status.
 */
  status?: 'CLOSED' | 'READY' | 'RECEIVED' | 'REJECTED' | 'SENT_FOR_REPAIR';
}

export function warrantyClaimsList(http: HttpClient, rootUrl: string, params?: WarrantyClaimsList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedWarrantyClaimList>> {
  const rb = new RequestBuilder(rootUrl, warrantyClaimsList.PATH, 'get');
  if (params) {
    rb.query('open', params.open, {});
    rb.query('out_of_warranty', params.out_of_warranty, {});
    rb.query('page', params.page, {});
    rb.query('search', params.search, {});
    rb.query('status', params.status, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedWarrantyClaimList>;
    })
  );
}

warrantyClaimsList.PATH = '/api/warranty/claims/';