// src/app/api/models/claim-status-enum.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * * `RECEIVED` - Received
 * * `SENT_FOR_REPAIR` - Sent for repair
 * * `READY` - Ready for collection
 * * `CLOSED` - Closed
 * * `REJECTED` - Rejected
 */
export type ClaimStatusEnum = 'RECEIVED' | 'SENT_FOR_REPAIR' | 'READY' | 'CLOSED' | 'REJECTED';